Refactor QuranReader and StreakGoalCard components for improved functionality
- Removed unused imports and code related to country language preferences.
- Simplified StreakGoalCard by integrating GuestStateCard for better handling of user states.
- Added utility function for intelligent subtitle formatting in StreakGoalCard.
- Updated imports in various components to ensure consistency and clarity.
