+++ src/components/HomePage/ReadingSection/StreakOrGoalCard/index.tsx
@@ -29,8 +29,6 @@ type Props = {
 
 const StreakOrGoalCard: React.FC<Props> = ({ goal, streak, currentActivityDay }) => {
   const { t, lang } = useTranslation('home');
-  const percent = goal?.isCompleted ? 100 : Math.min(goal?.progress?.percent || 0, 100);
-  const localizedPercent = toLocalizedNumber(percent, lang);
 
   const onStreakCardClicked = () => {
     logButtonClick('homepage_reading_streak_card');
