+++ src/components/QuranReader/EndOfSurahSection/StreakGoalCard/index.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import classNames from 'classnames';
 import useTranslation from 'next-translate/useTranslation';
 
+import GuestStateCard from './GuestStateCard';
 import styles from './StreakGoalCard.module.scss';
 
 import Card from '@/components/HomePage/Card';
@@ -30,76 +31,14 @@ const StreakGoalCard: React.FC<StreakGoalCardProps> = ({ cardClassName }) => {
   const { t, lang } = useTranslation('quran-reader');
   const { goal, streak, currentActivityDay } = useGetStreakWithMetadata({ showDayName: true });
   const isMobile = useIsMobile();
-
   const displayStreak = streak > 0 ? streak : 0;
-  const hasGoalOrStreak = !!goal || displayStreak > 0;
 
-  const onGoalArrowClicked = () => {
-    logButtonClick('end_of_surah_goal_card_arrow');
-  };
-
-  const onSetGoalButtonClicked = () => {
-    logButtonClick('end_of_surah_goal_card_set_goal');
-  };
-
-  // Helper function to split subtitle text intelligently for line break
-  const getFormattedSubtitle = (text: string) => {
-    // Find the last comma in the text
-    const lastCommaIndex = text.lastIndexOf(',');
-
-    if (lastCommaIndex === -1) {
-      // No comma found, return text as is
-      return text;
-    }
-
-    // Split at the last comma
-    const firstPart = text.substring(0, lastCommaIndex + 1); // Include the comma
-    const secondPart = text.substring(lastCommaIndex + 1).trim();
-
-    return (
-      <>
-        {firstPart}
-        <br />
-        {secondPart}
-      </>
-    );
-  };
-
-  if (!hasGoalOrStreak) {
-    return (
-      <Card
-        className={classNames(styles.endOfSurahCard, styles.guestState, cardClassName)}
-        data-testid="streak-goal-card"
-      >
-        <div className={styles.container}>
-          <div className={styles.titleContainer}>
-            <CirclesIcon className={styles.titleIcon} />
-            <h3 className={styles.title}>
-              {isMobile
-                ? t('end-of-surah.track-your-journey')
-                : t('end-of-surah.achieve-quran-goals-responsive')}
-            </h3>
-          </div>
-          {!isMobile && (
-            <p className={styles.subtitle}>
-              {getFormattedSubtitle(t('end-of-surah.achieve-quran-goals-desktop'))}
-            </p>
-          )}
-          <Button
-            type={ButtonType.Success}
-            size={isMobile ? ButtonSize.Small : ButtonSize.Medium}
-            href={getReadingGoalNavigationUrl()}
-            className={styles.button}
-            onClick={onSetGoalButtonClicked}
-          >
-            <CirclesIcon className={styles.buttonIcon} />
-            {isMobile ? t('end-of-surah.set-goal-mobile') : t('end-of-surah.set-custom-goal')}
-          </Button>
-        </div>
-      </Card>
-    );
+  if (!goal && displayStreak === 0) {
+    return <GuestStateCard cardClassName={cardClassName} />;
   }
 
+  const onGoalArrowClicked = () => logButtonClick('end_of_surah_goal_card_arrow');
+  const onSetGoalButtonClicked = () => logButtonClick('end_of_surah_goal_card_set_goal');
   const streakLabel =
     displayStreak === 1 ? t('end-of-surah.day-streak') : t('end-of-surah.days-streak');
